route decorated lambdas moved into the routes file
